+++ changelog.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2022-12-9
+
+### Added
+- Add `TMPHyperlink` component.
+- Add `HasName` method to Yarn extensions.
+
 ## [1.1.1] - 2022-11-24
 
 ### Fixed